+++ CHANGELOG.md
@@ -13,6 +13,8 @@ It is based on Tidepool Blip 1.27.
 - YLP-205 Change menu item "print"
 - YLP-229 Make "Daily" view the homepage of YourLoops instead of "Overview"
 - YLP-111 Display localized CGU and data privacy
+### Fixed
+- Be sure to use the fallback language if the navigator one is not supported (YLP-111 related)
 
 ### Engineering Use
 - YLP-221: move to Jenkins and automatically deploy our master branch (dblp) to cloudfront preview env.